requirements.txt
scipy==1.2.1
numpy~=1.19.2
matplotlib==3.1.2
opencv_python==4.1.2.30
torch==1.8.0
torchvision==0.9.0
tqdm~=4.27.0
Pillow==8.2.0
h5py==2.10.0
torchsummary~=1.5.1
pyrealsense2~=2.50.0.3812